@@ -9,7 +9,7 @@ class StudentApplicationsController < ApplicationController
     @student_application = StudentApplication.new(student_application_params)
     @student_application.user = current_user
     @student_application.semester = Semester.current
-    if @student_application.save
+    if @student_application.save_and_update_user
       flash[:info] = "Thanks for applying! We'll get back to you with our decision."
       redirect_to why_path
     else
@@ -22,7 +22,7 @@ class StudentApplicationsController < ApplicationController
 
   def student_application_params
     params.require(:student_application).permit(
-      :first_name, :last_name, :email, :year,
+      :first_name, :last_name, :email, :year, :standing,
       :major, :why_join, :cs_classes_taken, :current_courseload,
       :other_commitments, :how_many_hours_willing, :how_did_you_hear_about_us
     )
