@@ -9,12 +9,13 @@ class Ability
 
     can :create, CheckIn
     can :view, Lecture
+    can :view, Assignment
 
     if user.is_current_staff?
       can :manage, :all
     elsif user.enabled?
       can :show, User, id: user.id
-    elsif !user.new_record? # signed in but not enabled
+    elsif !user.new_record?
       can :apply, StudentApplication if !user.submitted_current_semester_application?
     end
   end
