@@ -3,7 +3,7 @@
 # Table name: roles
 #
 #  id          :integer          not null, primary key
-#  name        :string(255)
+#  name        :string
 #  created_at  :datetime
 #  updated_at  :datetime
 #  user_id     :integer
