@@ -3,13 +3,13 @@
 # Table name: student_applications
 #
 #  id                        :integer          not null, primary key
-#  first_name                :string(255)
-#  last_name                 :string(255)
-#  email                     :string(255)
-#  phone_number              :string(255)
-#  year                      :string(255)
-#  major                     :string(255)
-#  gpa                       :string(255)
+#  first_name                :string
+#  last_name                 :string
+#  email                     :string
+#  phone_number              :string
+#  year                      :string
+#  major                     :string
+#  gpa                       :string
 #  why_join                  :text
 #  cs_classes_taken          :text
 #  current_courseload        :text
@@ -20,15 +20,21 @@
 #  updated_at                :datetime
 #  user_id                   :integer
 #  semester_id               :integer
+#  standing                  :integer          default(0)
+#  status                    :integer          default(0)
 #
 
 class StudentApplication < ActiveRecord::Base
+  include StandingEnum
+
+  enum status: [ :pending, :accepted, :rejected ]
+
   belongs_to :user
   belongs_to :semester
 
   has_many :evaluations, dependent: :destroy
 
-  validates_presence_of :first_name, :last_name, :email, :year,
+  validates_presence_of :first_name, :last_name, :email, :year, :standing,
                         :major, :why_join, :cs_classes_taken, :current_courseload,
                         :other_commitments, :how_many_hours_willing, :how_did_you_hear_about_us
 
@@ -42,6 +48,10 @@ class StudentApplication < ActiveRecord::Base
   def next
     # Returns nil if it cannot find the next application.
     StudentApplication.find_by_id(id + 1)
+  end
+
+  def save_and_update_user
+    save && user.update(standing: standing)
   end
 
   def evaluations_by_users(users)
@@ -69,4 +79,29 @@ class StudentApplication < ActiveRecord::Base
   def total_no
     total('no')
   end
+
+  def total_accepted_in_standing
+    semester.accepted_student_applications.send(standing).count
+  end
+
+  def limit_for_standing
+    if standing == "upperclassman"
+      semester.upper_division_limit
+    else
+      semester.lower_division_limit
+    end
+  end
+
+  def accept
+    if total_accepted_in_standing < limit_for_standing
+      accepted!
+    end
+    accepted?
+  end
+
+  def pend
+    pending!
+    pending?
+  end
+
 end
