@@ -1,6 +1,8 @@
 RailsDecal::Application.routes.draw do
   root to: "pages#home"
 
+  resources :assignments
+
   match 'why', to: 'pages#why_this_class', via: :get
   match 'office-hours', to: 'pages#office_hours', via: :get
   match "apply", to: "student_applications#new", as: "apply", via: :get
@@ -31,12 +33,19 @@ RailsDecal::Application.routes.draw do
         resources :student_applications, path: '', only: [:show, :index] do
           member do
             match 'evaluate', to: 'evaluations#create', as: 'evaluate', via: [:post, :patch]
+            patch 'accept', to: 'student_applications#accept', as: 'accept'
+            patch 'pend', to: 'student_applications#pend', as: 'pend'
           end
         end
       end
     end
 
-    resources :users, only: [:index]
+    resources :users, only: [:index] do
+      collection do
+        get 'students', to: 'users#students'
+      end
+    end
+
     resources :check_in_codes, path: "check-in-codes", only: [:index, :show, :new, :create] do
       member do
         match "disable", to: "check_in_codes#disable", via: :post
