@@ -11,12 +11,40 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140911062516) do
+ActiveRecord::Schema.define(version: 20150725024617) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "check_in_codes", force: true do |t|
+  create_table "assignment_submissions", force: :cascade do |t|
+    t.integer  "user_id"
+    t.integer  "assignment_id"
+    t.integer  "points"
+    t.datetime "time_submitted"
+    t.integer  "grader_id"
+    t.datetime "created_at",     null: false
+    t.datetime "updated_at",     null: false
+  end
+
+  add_index "assignment_submissions", ["assignment_id"], name: "index_assignment_submissions_on_assignment_id", using: :btree
+  add_index "assignment_submissions", ["grader_id"], name: "index_assignment_submissions_on_grader_id", using: :btree
+  add_index "assignment_submissions", ["user_id"], name: "index_assignment_submissions_on_user_id", using: :btree
+
+  create_table "assignments", force: :cascade do |t|
+    t.integer  "semester_id"
+    t.string   "link"
+    t.string   "name"
+    t.integer  "category",    default: 0
+    t.float    "weight",      default: 1.0
+    t.integer  "points"
+    t.datetime "deadline"
+    t.datetime "created_at",                null: false
+    t.datetime "updated_at",                null: false
+  end
+
+  add_index "assignments", ["semester_id"], name: "index_assignments_on_semester_id", using: :btree
+
+  create_table "check_in_codes", force: :cascade do |t|
     t.string   "code"
     t.date     "class_date"
     t.datetime "created_at"
@@ -24,14 +52,14 @@ ActiveRecord::Schema.define(version: 20140911062516) do
     t.boolean  "enabled",    default: false
   end
 
-  create_table "check_ins", force: true do |t|
+  create_table "check_ins", force: :cascade do |t|
     t.integer  "user_id"
     t.integer  "check_in_code_id"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "evaluations", force: true do |t|
+  create_table "evaluations", force: :cascade do |t|
     t.text     "decision"
     t.integer  "user_id"
     t.integer  "student_application_id"
@@ -43,7 +71,7 @@ ActiveRecord::Schema.define(version: 20140911062516) do
   add_index "evaluations", ["student_application_id"], name: "index_evaluations_on_student_application_id", using: :btree
   add_index "evaluations", ["user_id"], name: "index_evaluations_on_user_id", using: :btree
 
-  create_table "lectures", force: true do |t|
+  create_table "lectures", force: :cascade do |t|
     t.integer  "number"
     t.string   "title"
     t.string   "partial"
@@ -52,13 +80,13 @@ ActiveRecord::Schema.define(version: 20140911062516) do
     t.integer  "semester_id"
   end
 
-  create_table "positions", force: true do |t|
+  create_table "positions", force: :cascade do |t|
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string   "name"
   end
 
-  create_table "roles", force: true do |t|
+  create_table "roles", force: :cascade do |t|
     t.string   "name"
     t.datetime "created_at"
     t.datetime "updated_at"
@@ -67,15 +95,17 @@ ActiveRecord::Schema.define(version: 20140911062516) do
     t.integer  "position_id"
   end
 
-  create_table "semesters", force: true do |t|
+  create_table "semesters", force: :cascade do |t|
     t.string   "semester"
     t.integer  "year"
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string   "url"
+    t.integer  "lower_division_limit"
+    t.integer  "upper_division_limit"
   end
 
-  create_table "student_applications", force: true do |t|
+  create_table "student_applications", force: :cascade do |t|
     t.string   "first_name"
     t.string   "last_name"
     t.string   "email"
@@ -93,9 +123,11 @@ ActiveRecord::Schema.define(version: 20140911062516) do
     t.datetime "updated_at"
     t.integer  "user_id"
     t.integer  "semester_id"
+    t.integer  "standing",                  default: 0
+    t.integer  "status",                    default: 0
   end
 
-  create_table "users", force: true do |t|
+  create_table "users", force: :cascade do |t|
     t.string   "email",                  default: "",    null: false
     t.string   "encrypted_password",     default: "",    null: false
     t.string   "reset_password_token"
@@ -117,8 +149,12 @@ ActiveRecord::Schema.define(version: 20140911062516) do
     t.string   "blog"
     t.string   "location"
     t.boolean  "enabled",                default: false
+    t.integer  "standing",               default: 0
   end
 
   add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
 
+  add_foreign_key "assignment_submissions", "assignments"
+  add_foreign_key "assignment_submissions", "users"
+  add_foreign_key "assignments", "semesters"
 end
